+++ interfaces/cs/Socketron/Node/NodeJS.cs
@@ -56,7 +56,7 @@ namespace Socketron {
 				item.CallbackId,
 				Script.AddObject("callback")
 			);
-			long objectId = _ExecuteBlocking<long>(script);
+			int objectId = _ExecuteBlocking<int>(script);
 			item.ObjectId = objectId;
 
 			script = ScriptBuilder.Build(
@@ -103,7 +103,7 @@ namespace Socketron {
 				item.CallbackId,
 				Script.AddObject("callback")
 			);
-			long objectId = _ExecuteBlocking<long>(script);
+			int objectId = _ExecuteBlocking<int>(script);
 			item.ObjectId = objectId;
 
 			script = ScriptBuilder.Build(
@@ -153,7 +153,7 @@ namespace Socketron {
 				item.CallbackId,
 				Script.AddObject("callback")
 			);
-			long objectId = _ExecuteBlocking<long>(script);
+			int objectId = _ExecuteBlocking<int>(script);
 			item.ObjectId = objectId;
 
 			script = ScriptBuilder.Build(
